pyproject.toml
[project]
name = "gfi-website"
version = "0.1.0"
description = "General Food Industry Co., Ltd. B2B marketing website"
authors = [
    {name = "NufelRokni"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-jazzmin (>=3.0.4,<4.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "boto3 (>=1.42.84,<2.0.0)",
    "django-summernote (>=0.8.20,<0.9.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "django-browser-reload (>=1.21.0,<2.0.0)"
]
